+++ app/models/bento_search/result_item.rb
@@ -1,4 +1,7 @@
 require 'language_list'
+require 'bento_search/author'
+require 'bento_search/link'
+
 
 module BentoSearch
   # Data object representing a single hit from a search, normalized
@@ -13,6 +16,8 @@ module BentoSearch
   class ResultItem
     include ERB::Util # for html_escape for our presentational stuff
     include ActionView::Helpers::OutputSafetyHelper # for safe_join
+
+    include ::BentoSearch::Results::Serialization
 
     # Can initialize with a hash of key/values
     def initialize(args = {})
@@ -30,21 +35,24 @@ module BentoSearch
     # internal unique id for the document, from the particular
     # search service it came from. May be alphanumeric. May be nil
     # for engines that don't support it.
-    attr_accessor :unique_id
+    serializable_attr_accessor :unique_id
+    
 
     # If set to true, item will refuse to generate an openurl,
     # returning nil from #to_openurl or #openurl_kev
-    attr_accessor :openurl_disabled
+    serializable_attr_accessor :openurl_disabled
+    
 
     # Array (possibly empty) of BentoSearch::Link objects
     # representing additional links. Often SearchEngine's themselves
     # won't include any of these, but Decorators will be used
     # to add them in.
     attr_accessor :other_links
+    serializable_attr :other_links, :collection_of => "BentoSearch::Link"
 
     # * dc.title
     # * schema.org CreativeWork: 'name'
-    attr_accessor :title
+    serializable_attr_accessor :title    
     # backwards compat, we used to have separate titles and subtitles
     alias_method :complete_title, :title
 
@@ -52,6 +60,7 @@ module BentoSearch
     # Can be changed in actual presentation with a Decorator.
     # * schema.org CreativeWork: 'url'
     attr_accessor :link
+    serializable_attr :link
 
     # does the #link correspond to fulltext?  true or false -- or nil
     # for unknown/non-applicable. Not all engines will set.
@@ -61,6 +70,7 @@ module BentoSearch
     def link_is_fulltext=(v)
       @link_is_fulltext = v
     end
+    serializable_attr :link_is_fulltext
 
     # Our own INTERNAL controlled vocab for 'format'.
     #
@@ -102,7 +112,7 @@ module BentoSearch
     #
     # Note: We're re-thinking this, might allow uncontrolled
     # in here instead.
-    attr_accessor :format
+    serializable_attr_accessor :format    
 
     # Translated from internal format vocab at #format. Outputs
     # eg http://schema.org/Book
@@ -127,7 +137,7 @@ module BentoSearch
     # uncontrolled presumably english-language format string.
     # if supplied will be used in display in place of controlled
     # format.
-    attr_accessor :format_str
+    serializable_attr_accessor :format_str    
 
     # Language of materials. Producer can set language_code to an ISO 639-1 (two
     # letter) or 639-3 (three letter) language code. If you do this, you don't
@@ -143,7 +153,7 @@ module BentoSearch
     # #language_iso_639_2 (either may be null), or #language_str for uncontrolled
     # string. If engine just sets one of these, internals take care of filling
     # out the others. r
-    attr_accessor :language_code
+    serializable_attr_accessor :language_code    
     attr_writer :language_str
     def language_str
       @language_str || language_code.try do |code|
@@ -152,6 +162,7 @@ module BentoSearch
         end
       end
     end
+    serializable_attr :language_str
     # Returns a LanguageList gem language object-- from #language_code
     # if available, otherwise from direct language_str if available and
     # possible.
@@ -176,45 +187,46 @@ module BentoSearch
     # * prism:coverDate, year portion
     #
     # See also publication_date when you have a complete date
-    attr_accessor :year
+    serializable_attr_accessor :year
     # ruby stdlib Date object.
     attr_accessor :publication_date
+    serializable_attr :publication_date, :serializer => "Date"
 
-    attr_accessor :volume
-    attr_accessor :issue
-    attr_accessor :start_page
-    attr_accessor :end_page
+    serializable_attr_accessor :volume
+    serializable_attr_accessor :issue
+    serializable_attr_accessor :start_page
+    serializable_attr_accessor :end_page
 
     # source_title is often used for journal_title (and aliased
     # as #journal_title, although that may go away), but can
     # also be used for other 'container' titles. Book title for
     # a book chapter. Even web site or URL for a web page.
-    attr_accessor :source_title
+    serializable_attr_accessor :source_title
     alias_method :journal_title, :source_title
     alias_method :'journal_title=',  :'source_title='
 
 
-    attr_accessor :issn
-    attr_accessor :isbn
-    attr_accessor :oclcnum # OCLC accession number, WorldCat.
+    serializable_attr_accessor :issn
+    serializable_attr_accessor :isbn
+    serializable_attr_accessor :oclcnum # OCLC accession number, WorldCat.
 
-    attr_accessor :doi
-    attr_accessor :pmid
+    serializable_attr_accessor :doi
+    serializable_attr_accessor :pmid
 
     # usually used for books rather than articles
-    attr_accessor :publisher
+    serializable_attr_accessor :publisher
 
     # an openurl kev-encoded context object. optional,
     # only if source provides one that may be better
     # than can be constructed from individual elements above
-    attr_accessor :openurl_kev_co
+    serializable_attr_accessor :openurl_kev_co
 
     # Short summary of item.
     # Mark .html_safe if it includes html -- creator is responsible
     # for making sure html is safely sanitizied and/or stripped,
     # rails ActionView::Helpers::Sanistize #sanitize and #strip_tags
     # may be helpful.
-    attr_accessor :abstract
+    serializable_attr_accessor :abstract
 
     # An ARRAY of string query-in-context snippets. Will usually
     # have highlighting <b> tags in it. Creator is responsible
@@ -225,14 +237,16 @@ module BentoSearch
     # with same content formatted differently (array of multiple vs
     # one combined string), some engines they may be different. 
     attr_accessor :snippets
+    serializable_attr :snippets
 
     # An array (order matters) of BentoSearch::Author objects
     # add authors to it with results.authors << Author
     attr_accessor :authors
+    serializable_attr :authors, :collection_of => "BentoSearch::Author"
 
     # engine-specific data not suitable for abstract API, usually
     # for internal use.
-    attr_accessor :custom_data
+    serializable_attr_accessor :custom_data
 
     # Copied over from engine configuration usually, a string
     # qualified name of a decorator class. Can be nil for default.
@@ -248,6 +262,8 @@ module BentoSearch
     # into a backpointing reference instead? And user cover-methods
     # for it? Nice thing about the configuration has instead is it's
     # easily serializable, it's just data.
+    #
+    # Although we intentionally do NOT include these in JSON serialization, ha. 
     attr_accessor :display_configuration
     attr_accessor :engine_id